pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "openai>=1.82.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot==20.3",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "telegram>=0.0.1",
]
